+++ Classes/VTAcknowledgementsViewController.h
@@ -77,17 +77,6 @@
 + (nonnull NSString *)localizedTitle;
 
 /**
- **Deprecated** Initializes an acknowledgements view controller with the content of an acknowledgements file (by its path).
-
- @param acknowledgementsPlistPath The path to the acknowledgements `.plist` file.
-
- @return A newly created `VTAcknowledgementsViewController` instance.
-
- @see -initWithPlistPath:
- */
-- (nullable instancetype)initWithAcknowledgementsPlistPath:(nullable NSString *)acknowledgementsPlistPath DEPRECATED_MSG_ATTRIBUTE("use -initWithPath: method instead");
-
-/**
  Initializes an acknowledgements view controller with the content of an acknowledgements file (by its path).
 
  @param acknowledgementsPlistPath The path to the acknowledgements `.plist` file.
@@ -95,17 +84,6 @@
  @return A newly created `VTAcknowledgementsViewController` instance.
  */
 - (nullable instancetype)initWithPath:(nullable NSString *)acknowledgementsPlistPath;
-
-/**
- **Deprecated** Initializes an acknowledgements view controller with the content of an acknowledgements file (by its name).
-
- @param acknowledgementsFileName The file name for the acknowledgements `.plist` file from the main bundle.
-
- @return A newly created `VTAcknowledgementsViewController` instance.
-
- @see -initWithFileNamed:
- */
-- (nullable instancetype)initWithAcknowledgementsFileNamed:(nullable NSString *)acknowledgementsFileName DEPRECATED_MSG_ATTRIBUTE("use -initWithFileNamed: method instead");
 
 /**
  Initializes an acknowledgements view controller with the content of an acknowledgements file (by its name).